blame tests: don't rely on t/t4018/ directory
Refactor a test added in 9466e3809d (blame: enable funcname blaming
with userdiff driver, 2020-11-01) so that the blame tests don't rely
on stealing the contents of "t/t4018/fortran-external-function".

I'm about to refactor that directory to delete that file, just moving
the relevant test file here inline is the easiest solution, and I
think also the most readable.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>