fsck.c: remove redundant sanity check in fsck_walk_tree()
Remove a check for whether tree modes are bad in
fsck_walk_tree(). This check duplicates logic we get from fsck_tree()
via fsck_object().

When this code was added in 355885d531 (add generic, type aware object
chain walker, 2008-02-25) we'd already had a better version of this
check in since 42ea9cb286 (Be more careful about tree entry modes.,
2005-05-05).

The same series that added the check moved fsck_object() into fsck.c
in ba002f3b28 (builtin-fsck: move common object checking code to
fsck.c, 2008-02-25).

There's no reason for us to be conflating fsck error reporting and
tree walking like this. Everything that's using the walking and cares
about correctness is already using fsck_object().

In terms of API sanity (not that I have a current use-case for this)
it makes sense for the walking API to be very forgiving, and the
fsck_object() function to be strict. We'd like to optionally enable
iterating over bad repository data, but shouldn't die early in case we
need to get more entries to walk down.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>